Move remapParameterValuesToTemplateTags to where it is used
