+++ frontend/src/metabase-lib/v1/parameters/utils/template-tags.ts
@@ -4,7 +4,6 @@ import * as Lib from "metabase-lib";
 import Question from "metabase-lib/v1/Question";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
 import type { ParameterWithTarget } from "metabase-lib/v1/parameters/types";
-import { getTemplateTagFromTarget } from "metabase-lib/v1/parameters/utils/targets";
 import { InternalQuery } from "metabase-lib/v1/queries/InternalQuery";
 import type {
   Card,
@@ -120,35 +119,4 @@ export function getTemplateTagParametersFromCard(
 ) {
   const tags = getTemplateTags(card, metadata);
   return getTemplateTagParameters(tags, card.parameters);
-}
-
-// when navigating from dashboard --> saved native question,
-// we are given dashboard parameters and a map of dashboard parameter ids to parameter values
-// we need to transform this into a map of template tag ids to parameter values
-// so that we populate the template tags in the native editor
-export function remapParameterValuesToTemplateTags(
-  templateTags: TemplateTag[],
-  dashboardParameters: ParameterWithTarget[],
-  parameterValuesByDashboardParameterId: {
-    [key: string]: any;
-  },
-) {
-  const parameterValues: {
-    [key: string]: any;
-  } = {};
-  const templateTagParametersByName = _.indexBy(templateTags, "name");
-
-  dashboardParameters.forEach((dashboardParameter) => {
-    const { target } = dashboardParameter;
-    const tag = getTemplateTagFromTarget(target);
-
-    if (tag != null && templateTagParametersByName[tag]) {
-      const templateTagParameter = templateTagParametersByName[tag];
-      const parameterValue =
-        parameterValuesByDashboardParameterId[dashboardParameter.id];
-      parameterValues[templateTagParameter.name] = parameterValue;
-    }
-  });
-
-  return parameterValues;
 }
